chore:  Enforces KMS key tagging with valid environment values
Refactors the KMS tag enforcement policy to utilize `templatefile` for dynamic environment tag validation.
This change replaces the static `valid_environments` variable with a dynamic `environment_tags` variable,
allowing for flexible configuration of allowed environment tag values. The `tag-enforcement.json` policy
is updated to use `${environment_tags}` within the `ForAnyValue:StringNotLike` condition, ensuring that
KMS keys are tagged with valid environment values.

The main.tf file is updated to use the `templatefile` function to render the `tag-enforcement.json` policy
with the `environment_tags` variable. This ensures that the policy is dynamically generated based on the
configured environment tags.

A new `environment_tags` variable is added to `variables.tf` to define the valid environment tag values.