chore: Adds local backend and updates terraform AWS provider version
Adds backend configuration files for dev and prod environments to store the terraform state locally.

Updates the AWS provider version in the terraform configuration to allow versions greater than or equal to 1.10.

Introduces a new environment variable to specify the environment name. This allows for dynamic configuration based on the target environment.