feat[org]: Enables creation of AWS organization accounts and organizational units
This commit introduces the ability to create AWS organization accounts and organizational units. It replaces the data source for the organization with a resource and adds resources for creating accounts. It also defines new variables for account lists and adds outputs for organization IDs and accounts. The SCP modules are commented out.